+++ includes/ObserverPattern.h
@@ -17,5 +17,32 @@ namespace ObserverPattern{
 			virtual void updateData(const string sKey, void *pData) = 0;
 	};	// End of class myAbstractObserver
 
+	// Description:
+	// This abstract class is to send data to all observers. If you want
+	// use this class, you must define a new class to extand this class, 
+	// and then implement the abstract method `notifyData()`.
+	class mySubjectInterface{
+		// protected fields
+		protected:
+			// The list is to store all observer.
+			vector<myAbstractObserver*> m_poObserver;
+
+		public:
+			mySubjectInterface();
+			~mySubjectInterface();
+
+			// Description: Register observer into the list
+			// @param poObserver: The object pointer of the class myAbstractObserver
+			void registerObserver(myAbstractObserver *poObserver);
+
+			// Description: Remove the observer from the list
+			// @param poObserver: The object pointer of the class myAbstractObserver
+			void removeObserver(myAbstractObserver *poObserver);
+
+			// Description: Notify the data to observers
+			// @param sKey: The keyword of the data
+			// @param pData: The data will notify to observer.
+			virtual void notifyData(const string sKey, void *pData) = 0;
+	};	// End of class mySubjectInterface
 }	// End of namespace ObserverPattern