+++ ChangeLog.md
@@ -5,6 +5,7 @@ This document is created to record all changes.
 ## [Unrelease]
 ### Added
 - Add the header file `includes/ObserverPattern.h` to define the class `myAbstractObserver` for observer pattern.
+- Define the class `mySubjectInterface` for observer pattern.
 
 ### Changed